server/agent/requirements.txt
httpx>=0.27
google-genai>=1.0
feedparser>=6.0
pyyaml>=6.0
python-dotenv>=1.0
